refactor: パスのタイプインポートを変更
